feat: migrate full root layout to single JIVE tree
- LayoutModel::makeRootLayout: one ValueTree for the whole window
  (main-area with header/plugin/content-row + status bar); the
  controls/keyboard split uses flex-grow 1:1.5 with min/max heights,
  reproducing the native dynamic allocation declaratively
- MainComponent::resized() reduced to 3 lines (JIVE FlexBox owns layout)
- setPluginPanelExpanded updates the plugin-panel node height property
- Extract JIVE accessors into MainComponentJiveAccessors.cpp (#included),
  matching the Phase 11 architecture; MainComponent.cpp ~1580 -> ~930 lines
- Extend StyleCatalogTest with root layout interpretation + bounds checks
- Update architecture.md and current-iteration.md (Phase 11c/d complete)
