+++ source/MainComponent.cpp
@@ -195,7 +195,7 @@ void MainComponent::initialiseUi() {
     setLookAndFeel(lookAndFeel.get());
     setWantsKeyboardFocus(true);
 
-    // ── JIVE header bar ──
+    // ── JIVE root layout (header, plugin, controls, keyboard, status bar) ──
     {
         // Global style rules (loaded once; re-loading is idempotent).
         const auto styleFile
@@ -206,33 +206,21 @@ void MainComponent::initialiseUi() {
         }
 
         jiveInterpreter = std::make_unique<::jive::Interpreter>();
-        jiveInterpreter->getComponentFactory().set("SettingsButton", [] {
+        auto& factory = jiveInterpreter->getComponentFactory();
+
+        factory.set("SettingsButton", [] {
             auto btn = std::make_unique<juce::DrawableButton>("settings", juce::DrawableButton::ImageFitted);
             btn->setImages(createGearIcon(devpiano::jive::DesignTokens::get().textSecondary()).get(),
                            createGearIcon(devpiano::jive::DesignTokens::get().primary()).get(), nullptr);
             return btn;
         });
-
-        auto headerTree = devpiano::ui::jive::makeHeaderTree();
-        devpiano::ui::jive::StyleCatalog::get().applyToTree(headerTree);
-        jiveHeaderItem = jiveInterpreter->interpret(headerTree);
-        if (jiveHeaderItem != nullptr) {
-            addAndMakeVisible(jiveHeaderItem->getComponent().get());
-            if (auto* settingsItem = jive::findItemWithID(*jiveHeaderItem, "settings-btn"))
-                if (auto* settingsBtn = dynamic_cast<juce::Button*>(settingsItem->getComponent().get()))
-                    settingsBtn->onClick = [this] { showSettingsDialog(); };
-        }
-    }
-
-    // ── JIVE plugin panel ──
-    {
-        jiveInterpreter->getComponentFactory().set("PathEditor", [] {
+        factory.set("PathEditor", [] {
             auto editor = std::make_unique<juce::TextEditor>();
             editor->setMultiLine(false);
             editor->setReturnKeyStartsNewLine(false);
             return editor;
         });
-        jiveInterpreter->getComponentFactory().set("ListEditor", [] {
+        factory.set("ListEditor", [] {
             auto editor = std::make_unique<juce::TextEditor>();
             editor->setMultiLine(true);
             editor->setReadOnly(true);
@@ -243,55 +231,7 @@ void MainComponent::initialiseUi() {
             editor->setMouseClickGrabsKeyboardFocus(false);
             return editor;
         });
-
-        auto pluginTree = devpiano::ui::jive::makePluginPanelTree();
-        devpiano::ui::jive::StyleCatalog::get().applyToTree(pluginTree);
-        jivePluginPanelItem = jiveInterpreter->interpret(pluginTree);
-        if (jivePluginPanelItem != nullptr) {
-            addAndMakeVisible(jivePluginPanelItem->getComponent().get());
-
-            const auto wireButton = [this](const char* id, const std::function<void()>& action) {
-                if (auto* item = jive::findItemWithID(*jivePluginPanelItem, id))
-                    if (auto* btn = dynamic_cast<juce::Button*>(item->getComponent().get()))
-                        btn->onClick = action;
-            };
-            wireButton("load-btn", [this] { pluginOperationController->loadSelectedPlugin(); });
-            wireButton("unload-btn", [this] { pluginOperationController->unloadCurrentPlugin(); });
-            wireButton("editor-btn", [this] { pluginOperationController->togglePluginEditor(); });
-            wireButton("toggle-btn", [this] { setPluginPanelExpanded(!appSettings.pluginPanelExpanded); });
-            wireButton("scan-btn", [this] { pluginOperationController->scanPlugins(); });
-            wireButton("browse-btn", [this] { showPluginBrowseDialog(); });
-
-            if (auto* item = jive::findItemWithID(*jivePluginPanelItem, "plugin-selector"))
-                if (auto* combo = dynamic_cast<juce::ComboBox*>(item->getComponent().get())) {
-                    combo->setTextWhenNothingSelected(TRANS("Select a scanned plugin..."));
-                    combo->setWantsKeyboardFocus(false);
-                    combo->onChange = [this, combo] {
-                        if (combo->getSelectedItemIndex() >= 0)
-                            pluginOperationController->loadSelectedPlugin();
-                    };
-                }
-
-            if (auto* item = jive::findItemWithID(*jivePluginPanelItem, "plugin-filter-combo"))
-                if (auto* combo = dynamic_cast<juce::ComboBox*>(item->getComponent().get())) {
-                    combo->setWantsKeyboardFocus(false);
-                    combo->onChange = [this] { refreshPluginUiState(); };
-                }
-
-            if (auto* item = jive::findItemWithID(*jivePluginPanelItem, "plugin-path-editor"))
-                if (auto* editor = dynamic_cast<juce::TextEditor*>(item->getComponent().get()))
-                    editor->onReturnKey = [this] { pluginOperationController->scanPlugins(); };
-        }
-    }
-
-    const auto pluginRecovery = getPluginRecoverySettingsWithFallback();
-    setPluginPathText(makeSafeUiText(pluginRecovery.pluginSearchPath));
-
-    setPluginPanelExpanded(appSettings.pluginPanelExpanded);
-
-    // ── JIVE controls panel ──
-    {
-        jiveInterpreter->getComponentFactory().set("DevKnob", [] {
+        factory.set("DevKnob", [] {
             auto slider = std::make_unique<juce::Slider>();
             slider->setSliderStyle(juce::Slider::RotaryHorizontalVerticalDrag);
             slider->setTextBoxStyle(juce::Slider::TextBoxBelow, false, 44, 16);
@@ -299,10 +239,10 @@ void MainComponent::initialiseUi() {
                                         true);
             return slider;
         });
-        jiveInterpreter->getComponentFactory().set("AdsrCurve", [] { return std::make_unique<AdsrCurveComponent>(); });
+        factory.set("AdsrCurve", [] { return std::make_unique<AdsrCurveComponent>(); });
         const auto registerIconButton
-            = [this](const char* type, const std::unique_ptr<juce::Drawable>& image, const juce::String& tooltip) {
-                  jiveInterpreter->getComponentFactory().set(type, [&image, tooltip] {
+            = [&factory](const char* type, const std::unique_ptr<juce::Drawable>& image, const juce::String& tooltip) {
+                  factory.set(type, [&image, tooltip] {
                       auto btn = std::make_unique<juce::DrawableButton>(tooltip, juce::DrawableButton::ImageFitted);
                       btn->setImages(image.get());
                       btn->setTooltip(tooltip);
@@ -318,21 +258,18 @@ void MainComponent::initialiseUi() {
         registerIconButton("PlayButton", playIcon, TRANS("Play"));
         registerIconButton("StopButton", stopIcon, TRANS("Stop"));
         registerIconButton("BackButton", backIcon, TRANS("Back to Start"));
+        factory.set("CustomKeyboard",
+                    [this] { return std::make_unique<KeyboardViewport>(audioEngine.getKeyboardState()); });
+        factory.set("StatusBarMidiDot", [] { return std::make_unique<StatusBarMidiDot>(); });
 
-        auto controlsTree = devpiano::ui::jive::makeControlsPanelTree();
-        devpiano::ui::jive::StyleCatalog::get().applyToTree(controlsTree);
-        jiveControlsPanelItem = jiveInterpreter->interpret(controlsTree);
-        if (jiveControlsPanelItem != nullptr) {
-            addAndMakeVisible(jiveControlsPanelItem->getComponent().get());
+        auto rootTree = devpiano::ui::jive::makeRootLayout();
+        devpiano::ui::jive::StyleCatalog::get().applyToTree(rootTree);
+        jiveRootItem = jiveInterpreter->interpret(rootTree);
+        if (jiveRootItem != nullptr) {
+            addAndMakeVisible(jiveRootItem->getComponent().get());
 
-            auto* adsrCurve = []() -> AdsrCurveComponent* {
-                // resolved after wiring below
-                return nullptr;
-            }();
-
-            const auto findItem = [this](const char* id) -> ::jive::GuiItem* {
-                return jive::findItemWithID(*jiveControlsPanelItem, id);
-            };
+            const auto findItem
+                = [this](const char* id) -> ::jive::GuiItem* { return jive::findItemWithID(*jiveRootItem, id); };
             const auto findSlider = [&findItem](const char* id) -> juce::Slider* {
                 if (auto* item = findItem(id))
                     return dynamic_cast<juce::Slider*>(item->getComponent().get());
@@ -348,12 +285,46 @@ void MainComponent::initialiseUi() {
                     return dynamic_cast<juce::Button*>(item->getComponent().get());
                 return nullptr;
             };
-            const auto setButtonEnabled = [this](const char* id, bool enabled) {
-                if (auto* item = jive::findItemWithID(*jiveControlsPanelItem, id))
+            const auto setEnabled = [&findItem](const char* id, bool enabled) {
+                if (auto* item = findItem(id))
                     item->state.setProperty("enabled", enabled, nullptr);
             };
 
-            if (auto* item = jive::findItemWithID(*jiveControlsPanelItem, "adsr-curve"))
+            // ── header ──
+            if (auto* btn = findButton("settings-btn"))
+                btn->onClick = [this] { showSettingsDialog(); };
+
+            // ── plugin panel ──
+            const auto wireButton = [&findButton](const char* id, const std::function<void()>& action) {
+                if (auto* btn = findButton(id))
+                    btn->onClick = action;
+            };
+            wireButton("load-btn", [this] { pluginOperationController->loadSelectedPlugin(); });
+            wireButton("unload-btn", [this] { pluginOperationController->unloadCurrentPlugin(); });
+            wireButton("editor-btn", [this] { pluginOperationController->togglePluginEditor(); });
+            wireButton("toggle-btn", [this] { setPluginPanelExpanded(!appSettings.pluginPanelExpanded); });
+            wireButton("scan-btn", [this] { pluginOperationController->scanPlugins(); });
+            wireButton("browse-btn", [this] { showPluginBrowseDialog(); });
+
+            if (auto* combo = findCombo("plugin-selector")) {
+                combo->setTextWhenNothingSelected(TRANS("Select a scanned plugin..."));
+                combo->setWantsKeyboardFocus(false);
+                combo->onChange = [this, combo] {
+                    if (combo->getSelectedItemIndex() >= 0)
+                        pluginOperationController->loadSelectedPlugin();
+                };
+            }
+            if (auto* combo = findCombo("plugin-filter-combo")) {
+                combo->setWantsKeyboardFocus(false);
+                combo->onChange = [this] { refreshPluginUiState(); };
+            }
+            if (auto* item = findItem("plugin-path-editor"))
+                if (auto* editor = dynamic_cast<juce::TextEditor*>(item->getComponent().get()))
+                    editor->onReturnKey = [this] { pluginOperationController->scanPlugins(); };
+
+            // ── controls panel ──
+            auto* adsrCurve = []() -> AdsrCurveComponent* { return nullptr; }();
+            if (auto* item = findItem("adsr-curve"))
                 adsrCurve = dynamic_cast<AdsrCurveComponent*>(item->getComponent().get());
 
             const auto wireKnob = [&findSlider](const char* id, double min, double max, double interval,
@@ -365,7 +336,6 @@ void MainComponent::initialiseUi() {
                     slider->onValueChange = [onChanged] { onChanged(); };
                 }
             };
-
             wireKnob(
                 "volume-knob", 0.0, 1.0, 0.01, [](double v) { return juce::String(v, 2); },
                 [this] { handlePerformanceUiChanged(); });
@@ -386,24 +356,20 @@ void MainComponent::initialiseUi() {
                 "speed-knob", 0.5, 2.0, 0.25, [](double v) { return juce::String(v, 1) + "x"; },
                 [this] { recordingSessionController->handlePlaybackSpeedChange(getControlsPlaybackSpeed()); });
 
-            const auto wireButtonAction = [this, &findButton](const char* id, const std::function<void()>& action) {
-                if (auto* btn = findButton(id))
-                    btn->onClick = action;
-            };
-            wireButtonAction("record-btn", [this] { recordingSessionController->handleRecordClicked(); });
-            wireButtonAction("play-btn", [this] { recordingSessionController->handlePlayClicked(); });
-            wireButtonAction("stop-btn", [this] { recordingSessionController->handleStopClicked(); });
-            wireButtonAction("back-btn", [this] { recordingSessionController->handleBackToStartClicked(); });
-            wireButtonAction("export-midi-btn", [this] { recordingSessionController->handleExportMidiClicked(); });
-            wireButtonAction("export-wav-btn", [this] { recordingSessionController->handleExportWavClicked(); });
-            wireButtonAction("import-midi-btn", [this] { recordingSessionController->handleImportMidiClicked(); });
-            wireButtonAction("save-perf-btn", [this] { recordingSessionController->handleSavePerformanceClicked(); });
-            wireButtonAction("open-perf-btn", [this] { recordingSessionController->handleOpenPerformanceClicked(); });
-            wireButtonAction("song-info-btn", [this] { recordingSessionController->handleSongInfoClicked(); });
-            wireButtonAction("recent-btn", [this] { showRecentFilesMenu(); });
-            wireButtonAction("save-preset-btn", [this] { presetFlowSupport->handleSaveAsNewPreset(); });
-            wireButtonAction("rename-preset-btn", [this] { presetFlowSupport->handleRenamePreset(); });
-            wireButtonAction("delete-preset-btn", [this] { presetFlowSupport->handleDeletePreset(); });
+            wireButton("record-btn", [this] { recordingSessionController->handleRecordClicked(); });
+            wireButton("play-btn", [this] { recordingSessionController->handlePlayClicked(); });
+            wireButton("stop-btn", [this] { recordingSessionController->handleStopClicked(); });
+            wireButton("back-btn", [this] { recordingSessionController->handleBackToStartClicked(); });
+            wireButton("export-midi-btn", [this] { recordingSessionController->handleExportMidiClicked(); });
+            wireButton("export-wav-btn", [this] { recordingSessionController->handleExportWavClicked(); });
+            wireButton("import-midi-btn", [this] { recordingSessionController->handleImportMidiClicked(); });
+            wireButton("save-perf-btn", [this] { recordingSessionController->handleSavePerformanceClicked(); });
+            wireButton("open-perf-btn", [this] { recordingSessionController->handleOpenPerformanceClicked(); });
+            wireButton("song-info-btn", [this] { recordingSessionController->handleSongInfoClicked(); });
+            wireButton("recent-btn", [this] { showRecentFilesMenu(); });
+            wireButton("save-preset-btn", [this] { presetFlowSupport->handleSaveAsNewPreset(); });
+            wireButton("rename-preset-btn", [this] { presetFlowSupport->handleRenamePreset(); });
+            wireButton("delete-preset-btn", [this] { presetFlowSupport->handleDeletePreset(); });
 
             if (auto* combo = findCombo("preset-combo")) {
                 combo->setTextWhenNothingSelected("Default");
@@ -417,8 +383,18 @@ void MainComponent::initialiseUi() {
             }
 
             setRecordingControlsState({});
+
+            // ── keyboard area ──
+            if (auto* item = findItem("custom-keyboard"))
+                if (auto* viewport = dynamic_cast<KeyboardViewport*>(item->getComponent().get()))
+                    customKeyboardRef = &viewport->getCustomKeyboard();
         }
     }
+
+    const auto pluginRecovery = getPluginRecoverySettingsWithFallback();
+    setPluginPathText(makeSafeUiText(pluginRecovery.pluginSearchPath));
+
+    setPluginPanelExpanded(appSettings.pluginPanelExpanded);
 
     recordingSessionController->onFileOpened = [this](const juce::File& file) {
         recentFiles.addFile(file);
@@ -431,34 +407,6 @@ void MainComponent::initialiseUi() {
     // Initialize playback speed to 1.0x (never persisted — default on every launch).
     setControlsPlaybackSpeed(1.0);
     recordingEngine.setPlaybackSpeedMultiplier(1.0);
-
-    // ── JIVE keyboard area ──
-    {
-        jiveInterpreter->getComponentFactory().set(
-            "CustomKeyboard", [this] { return std::make_unique<KeyboardViewport>(audioEngine.getKeyboardState()); });
-
-        auto keyboardTree = devpiano::ui::jive::makeKeyboardAreaTree();
-        devpiano::ui::jive::StyleCatalog::get().applyToTree(keyboardTree);
-        jiveKeyboardAreaItem = jiveInterpreter->interpret(keyboardTree);
-        if (jiveKeyboardAreaItem != nullptr) {
-            addAndMakeVisible(jiveKeyboardAreaItem->getComponent().get());
-            if (auto* item = jive::findItemWithID(*jiveKeyboardAreaItem, "custom-keyboard"))
-                if (auto* viewport = dynamic_cast<KeyboardViewport*>(item->getComponent().get()))
-                    customKeyboardRef = &viewport->getCustomKeyboard();
-        }
-    }
-
-    // ── JIVE status bar ──
-    {
-        jiveInterpreter->getComponentFactory().set("StatusBarMidiDot",
-                                                   [] { return std::make_unique<StatusBarMidiDot>(); });
-
-        auto statusTree = devpiano::ui::jive::makeStatusBarTree();
-        devpiano::ui::jive::StyleCatalog::get().applyToTree(statusTree);
-        jiveStatusBarItem = jiveInterpreter->interpret(statusTree);
-        if (jiveStatusBarItem != nullptr)
-            addAndMakeVisible(jiveStatusBarItem->getComponent().get());
-    }
 
     // Wire CustomKeyboard mouse interaction → sound (with MIDI matrix)
     auto& customKeyboard = getCustomKeyboard();
@@ -639,70 +587,24 @@ void MainComponent::paint(juce::Graphics& g) {
         g.drawRect(fb, 1.0f);
     };
 
-    const auto headerBounds
-        = (jiveHeaderItem != nullptr) ? jiveHeaderItem->getComponent()->getBounds() : juce::Rectangle<int> {};
-    drawPanelBorder(headerBounds);
-    const auto pluginBounds
-        = (jivePluginPanelItem != nullptr) ? jivePluginPanelItem->getComponent()->getBounds() : juce::Rectangle<int> {};
-    drawPanelBorder(pluginBounds);
-    const auto controlsBounds = (jiveControlsPanelItem != nullptr) ? jiveControlsPanelItem->getComponent()->getBounds()
-                                                                   : juce::Rectangle<int> {};
-    drawPanelBorder(controlsBounds);
-    const auto keyboardBounds = (jiveKeyboardAreaItem != nullptr) ? jiveKeyboardAreaItem->getComponent()->getBounds()
-                                                                  : juce::Rectangle<int> {};
-    drawPanelBorder(keyboardBounds);
+    const auto panelBounds = [this](const char* id) {
+        if (jiveRootItem == nullptr)
+            return juce::Rectangle<int> {};
+        if (auto* item = jive::findItemWithID(*jiveRootItem, id))
+            return item->getComponent()->getBounds();
+        return juce::Rectangle<int> {};
+    };
+    drawPanelBorder(panelBounds("header"));
+    drawPanelBorder(panelBounds("plugin-panel"));
+    drawPanelBorder(panelBounds("controls-panel"));
+    drawPanelBorder(panelBounds("keyboard-area"));
 }
 
 void MainComponent::resized() {
-    auto area = getLocalBounds();
-    if (jiveStatusBarItem != nullptr)
-        jiveStatusBarItem->getComponent()->setBounds(area.removeFromBottom(22));
-    else
-        area.removeFromBottom(22);
-
-    auto content = area.reduced(16);
-    if (jiveHeaderItem != nullptr)
-        jiveHeaderItem->getComponent()->setBounds(content.removeFromTop(36));
-    else
-        content.removeFromTop(36);
-    content.removeFromTop(10);
-
-    if (jivePluginPanelItem != nullptr)
-        jivePluginPanelItem->getComponent()->setBounds(
-            content.removeFromTop(appSettings.pluginPanelExpanded ? 160 : 40));
-    else
-        content.removeFromTop(40);
-    content.removeFromTop(12);
-
-    // ── Dynamic allocation between ControlsPanel and KeyboardPanel ──
-    constexpr int controlsBase = 174;
-    constexpr int keyboardMin = 90;
-    constexpr int keyboardMax = 200;
-    constexpr int baseline = controlsBase + keyboardMin;
-    constexpr float keyboardRatio = 0.6f;
-
-    int alloc = content.getHeight() - 8; // gap between controls & keyboard
-
-    int keyboardHeight = keyboardMin;
-    if (alloc > baseline) {
-        int extra = alloc - baseline;
-        keyboardHeight = keyboardMin + static_cast<int>(static_cast<float>(extra) * keyboardRatio);
-        keyboardHeight = juce::jmin(keyboardHeight, keyboardMax);
-    } else {
-        keyboardHeight = juce::jmax(keyboardMin, alloc - controlsBase);
-    }
-
-    int controlsHeight = alloc - keyboardHeight;
-
-    if (jiveControlsPanelItem != nullptr)
-        jiveControlsPanelItem->getComponent()->setBounds(content.removeFromTop(controlsHeight));
-    else
-        content.removeFromTop(controlsHeight);
-    content.removeFromTop(8);
-    if (jiveKeyboardAreaItem != nullptr)
-        jiveKeyboardAreaItem->getComponent()->setBounds(content.removeFromTop(keyboardHeight));
-    else
-        content.removeFromTop(keyboardHeight);
+    // The entire layout is a single JIVE FlexBox tree; resizing the root
+    // component propagates to every panel.
+    if (jiveRootItem != nullptr)
+        jiveRootItem->getComponent()->setBounds(getLocalBounds());
 }
 
 void MainComponent::paintOverChildren(juce::Graphics& g) {
@@ -1027,651 +929,5 @@ void MainComponent::refreshPluginUiState() {
     renderReadOnlyUiState(buildCurrentAppStateSnapshot());
 }
 
-// ── JIVE plugin panel accessors ────────────────────────────────────────────
-
-void MainComponent::setPluginPathText(const juce::String& text) {
-    if (jivePluginPanelItem == nullptr)
-        return;
-    if (auto* item = jive::findItemWithID(*jivePluginPanelItem, "plugin-path-editor"))
-        if (auto* editor = dynamic_cast<juce::TextEditor*>(item->getComponent().get()))
-            editor->setText(text, juce::dontSendNotification);
-}
-
-juce::String MainComponent::getPluginPathText() const {
-    if (jivePluginPanelItem == nullptr)
-        return {};
-    if (auto* item = jive::findItemWithID(*jivePluginPanelItem, "plugin-path-editor"))
-        if (auto* editor = dynamic_cast<juce::TextEditor*>(item->getComponent().get()))
-            return editor->getText();
-    return {};
-}
-
-juce::String MainComponent::getSelectedPluginName() const {
-    if (jivePluginPanelItem == nullptr)
-        return {};
-    if (auto* item = jive::findItemWithID(*jivePluginPanelItem, "plugin-selector"))
-        if (auto* combo = dynamic_cast<juce::ComboBox*>(item->getComponent().get()))
-            return combo->getText();
-    return {};
-}
-
-void MainComponent::setPluginPanelExpanded(bool expanded) {
-    appSettings.pluginPanelExpanded = expanded;
-    if (jivePluginPanelItem != nullptr) {
-        if (auto* item = jive::findItemWithID(*jivePluginPanelItem, "plugin-expanded-area"))
-            item->state.setProperty("height", expanded ? 112 : 0, nullptr);
-        resized();
-    }
-    settingsStore.scheduleSave(appSettings);
-}
-
-void MainComponent::setInstrumentFilterVisible(bool visible) {
-    if (jivePluginPanelItem == nullptr)
-        return;
-    if (auto* item = jive::findItemWithID(*jivePluginPanelItem, "plugin-filter-combo"))
-        item->state.setProperty("visibility", visible, nullptr);
-}
-
-void MainComponent::showPluginBrowseDialog() {
-    auto chooser = std::make_shared<juce::FileChooser>(TRANS("Select VST3 Plugin Folder"),
-                                                       juce::File(getPluginPathText()), "", true);
-    chooser->launchAsync(juce::FileBrowserComponent::openMode | juce::FileBrowserComponent::canSelectDirectories,
-                         [this, chooser](const juce::FileChooser& fc) {
-                             auto folder = fc.getResult();
-                             if (folder.exists()) {
-                                 setPluginPathText(folder.getFullPathName());
-                                 pluginOperationController->scanPlugins();
-                             }
-                         });
-}
-
-void MainComponent::updatePluginPanelState(const PluginPanelState& state) {
-    if (jivePluginPanelItem == nullptr)
-        return;
-
-    auto* selectorItem = jive::findItemWithID(*jivePluginPanelItem, "plugin-selector");
-    auto* statusItem = jive::findItemWithID(*jivePluginPanelItem, "plugin-status-label");
-    auto* listItem = jive::findItemWithID(*jivePluginPanelItem, "plugin-list-editor");
-    auto* filterItem = jive::findItemWithID(*jivePluginPanelItem, "plugin-filter-combo");
-
-    auto* selectorCombo
-        = selectorItem != nullptr ? dynamic_cast<juce::ComboBox*>(selectorItem->getComponent().get()) : nullptr;
-    auto* listEditor = listItem != nullptr ? dynamic_cast<juce::TextEditor*>(listItem->getComponent().get()) : nullptr;
-
-    const auto setEnabled = [](::jive::GuiItem* item, bool enabled) {
-        if (item != nullptr)
-            item->state.setProperty("enabled", enabled, nullptr);
-    };
-
-    if (state.isCurrentlyScanning) {
-        if (selectorItem != nullptr)
-            selectorItem->state.removeAllChildren(nullptr);
-        if (selectorCombo != nullptr)
-            selectorCombo->setTextWhenNothingSelected(TRANS("Scanning..."));
-        if (listEditor != nullptr) {
-            auto scanText = TRANS("Scanning VST3 plugins...") + "\n";
-            scanText << (state.scanningPluginName.isNotEmpty() ? state.scanningPluginName : TRANS("Preparing..."));
-            listEditor->setText(scanText, juce::dontSendNotification);
-        }
-        setEnabled(jive::findItemWithID(*jivePluginPanelItem, "scan-btn"), false);
-        setEnabled(jive::findItemWithID(*jivePluginPanelItem, "browse-btn"), false);
-        setEnabled(jive::findItemWithID(*jivePluginPanelItem, "load-btn"), false);
-    } else {
-        const auto& names = [&]() -> const juce::StringArray& {
-            const auto filterId = filterItem != nullptr ? filterItem->state["selected"].toString().getIntValue() : 1;
-            if (filterId == 2 && !state.instrumentPluginNames.isEmpty())
-                return state.instrumentPluginNames;
-            if (filterId == 3 && !state.effectPluginNames.isEmpty())
-                return state.effectPluginNames;
-            return state.availablePluginNames;
-        }();
-
-        if (selectorItem != nullptr) {
-            selectorItem->state.removeAllChildren(nullptr);
-            auto selectedIndex = -1;
-            auto index = 0;
-            for (const auto& name : names) {
-                auto option = juce::ValueTree("Option");
-                option.setProperty("text", name, nullptr);
-                selectorItem->state.addChild(option, index, nullptr);
-                if (name.equalsIgnoreCase(state.preferredSelection))
-                    selectedIndex = index;
-                ++index;
-            }
-            if (names.isEmpty())
-                selectorItem->state.setProperty("selected", -1, nullptr);
-            else if (selectedIndex >= 0)
-                selectorItem->state.setProperty("selected", selectedIndex, nullptr);
-            else
-                selectorItem->state.setProperty("selected", 0, nullptr);
-        }
-
-        if (selectorCombo != nullptr)
-            selectorCombo->setTextWhenNothingSelected(TRANS("Select a scanned plugin..."));
-        if (listEditor != nullptr)
-            listEditor->setText(TRANS(state.pluginListText), juce::dontSendNotification);
-
-        setEnabled(jive::findItemWithID(*jivePluginPanelItem, "scan-btn"), true);
-        setEnabled(jive::findItemWithID(*jivePluginPanelItem, "browse-btn"), true);
-        setEnabled(jive::findItemWithID(*jivePluginPanelItem, "load-btn"), !names.isEmpty());
-        setEnabled(jive::findItemWithID(*jivePluginPanelItem, "unload-btn"), state.hasLoadedPlugin);
-        setEnabled(jive::findItemWithID(*jivePluginPanelItem, "editor-btn"), state.hasLoadedPlugin);
-        if (auto* item = jive::findItemWithID(*jivePluginPanelItem, "plugin-path-editor"))
-            item->state.setProperty("enabled", true, nullptr);
-    }
-
-    // Status line: formats + scan summary + loaded plugin info.
-    auto text = TRANS(state.availableFormatsDescription);
-    if (state.supportsVst3)
-        text << TRANS(" [VST3 ready]");
-
-    if (state.isCurrentlyScanning) {
-        text << TRANS(" | Scanning: ") << state.scanningPluginName << "...";
-    } else {
-        auto summary = state.lastScanSummary;
-        if (summary.startsWith("VST3 scan complete: ") && !summary.contains("no plugins")) {
-            auto resultSuffix = (state.scanFailedCount > 0) ? TRANS(" failed (see log).") : TRANS(" failed.");
-            text << " | " << TRANS("VST3 scan complete: ") << juce::String(state.scanPluginCount)
-                 << TRANS(" plugin(s), ") << juce::String(state.scanFailedCount) << resultSuffix;
-        } else if (summary.startsWith("VST3 scan found no plugins; ")) {
-            text << " | " << TRANS("VST3 scan found no plugins: ") << juce::String(state.scanFailedCount)
-                 << TRANS(" failed (see log).");
-        } else if (summary.startsWith("Loaded cached plugin list: ")) {
-            text << " | " << TRANS("Loaded cached plugin list: ") << juce::String(state.scanPluginCount)
-                 << TRANS(" plugin(s).");
-        } else {
-            text << " | " << TRANS(summary);
-        }
-    }
-
-    if (state.hasLoadedPlugin) {
-        text << TRANS(" | Loaded: ") << state.currentPluginName;
-
-        if (state.isPrepared)
-            text << " @ " << juce::String(state.preparedSampleRate, 0) << " Hz / "
-                 << juce::String(state.preparedBlockSize);
-        else
-            text << TRANS(" [not prepared]");
-
-        if (state.isEditorOpen)
-            text << TRANS(" | Editor open");
-    } else if (state.lastLoadError.isNotEmpty() && state.lastLoadError != "No plugin load attempted yet.") {
-        text << TRANS(" | Load error: ") << state.lastLoadError;
-    } else if (state.lastPluginName.isNotEmpty()) {
-        text << TRANS(" | Last plugin: ") << state.lastPluginName;
-    }
-
-    if (statusItem != nullptr)
-        statusItem->state.setProperty("text", text, nullptr);
-}
-
-void MainComponent::refreshPluginPanelTexts() {
-    if (jivePluginPanelItem == nullptr)
-        return;
-
-    if (auto* item = jive::findItemWithID(*jivePluginPanelItem, "plugin-path-label"))
-        item->state.setProperty("text", TRANS("VST3 Path"), nullptr);
-    const auto setButtonText = [this](const char* id, const juce::String& text) {
-        if (auto* item = jive::findItemWithID(*jivePluginPanelItem, id))
-            item->state.setProperty("text", text, nullptr);
-    };
-    setButtonText("scan-btn", TRANS("Scan VST3"));
-    setButtonText("load-btn", TRANS("Load"));
-    setButtonText("unload-btn", TRANS("Unload"));
-    setButtonText("editor-btn", TRANS("Open Editor"));
-    if (auto* item = jive::findItemWithID(*jivePluginPanelItem, "plugin-filter-combo")) {
-        const juce::StringArray filterTexts { TRANS("All"), TRANS("Instruments Only"), TRANS("Effects Only") };
-        auto childIndex = 0;
-        for (auto child : item->state) {
-            if (childIndex < filterTexts.size())
-                child.setProperty("text", filterTexts[childIndex], nullptr);
-            ++childIndex;
-        }
-    }
-    if (auto* item = jive::findItemWithID(*jivePluginPanelItem, "plugin-selector"))
-        if (auto* combo = dynamic_cast<juce::ComboBox*>(item->getComponent().get()))
-            combo->setTextWhenNothingSelected(TRANS("Select a scanned plugin..."));
-
-    // Re-apply the last state to refresh status text (locale-dependent).
-    refreshPluginUiState();
-}
-
-// ── JIVE controls panel accessors ──────────────────────────────────────────
-
-float MainComponent::getMasterGain() const {
-    if (jiveControlsPanelItem == nullptr)
-        return 0.0f;
-    if (auto* item = jive::findItemWithID(*jiveControlsPanelItem, "volume-knob"))
-        if (auto* slider = dynamic_cast<juce::Slider*>(item->getComponent().get()))
-            return static_cast<float>(slider->getValue());
-    return 0.0f;
-}
-
-float MainComponent::getAttack() const {
-    if (jiveControlsPanelItem == nullptr)
-        return 0.0f;
-    if (auto* item = jive::findItemWithID(*jiveControlsPanelItem, "attack-knob"))
-        if (auto* slider = dynamic_cast<juce::Slider*>(item->getComponent().get()))
-            return static_cast<float>(slider->getValue());
-    return 0.0f;
-}
-
-float MainComponent::getDecay() const {
-    if (jiveControlsPanelItem == nullptr)
-        return 0.0f;
-    if (auto* item = jive::findItemWithID(*jiveControlsPanelItem, "decay-knob"))
-        if (auto* slider = dynamic_cast<juce::Slider*>(item->getComponent().get()))
-            return static_cast<float>(slider->getValue());
-    return 0.0f;
-}
-
-float MainComponent::getSustain() const {
-    if (jiveControlsPanelItem == nullptr)
-        return 0.0f;
-    if (auto* item = jive::findItemWithID(*jiveControlsPanelItem, "sustain-knob"))
-        if (auto* slider = dynamic_cast<juce::Slider*>(item->getComponent().get()))
-            return static_cast<float>(slider->getValue());
-    return 0.0f;
-}
-
-float MainComponent::getRelease() const {
-    if (jiveControlsPanelItem == nullptr)
-        return 0.0f;
-    if (auto* item = jive::findItemWithID(*jiveControlsPanelItem, "release-knob"))
-        if (auto* slider = dynamic_cast<juce::Slider*>(item->getComponent().get()))
-            return static_cast<float>(slider->getValue());
-    return 0.0f;
-}
-
-double MainComponent::getControlsPlaybackSpeed() const {
-    if (jiveControlsPanelItem == nullptr)
-        return 1.0;
-    if (auto* item = jive::findItemWithID(*jiveControlsPanelItem, "speed-knob"))
-        if (auto* slider = dynamic_cast<juce::Slider*>(item->getComponent().get()))
-            return slider->getValue();
-    return 1.0;
-}
-
-void MainComponent::setControlsValues(float masterGain, float attack, float decay, float sustain, float release) {
-    if (jiveControlsPanelItem == nullptr)
-        return;
-    const auto setSlider = [this](const char* id, double value) {
-        if (auto* item = jive::findItemWithID(*jiveControlsPanelItem, id))
-            if (auto* slider = dynamic_cast<juce::Slider*>(item->getComponent().get()))
-                slider->setValue(value, juce::dontSendNotification);
-    };
-    setSlider("volume-knob", masterGain);
-    setSlider("attack-knob", attack);
-    setSlider("decay-knob", decay);
-    setSlider("sustain-knob", sustain);
-    setSlider("release-knob", release);
-    if (auto* item = jive::findItemWithID(*jiveControlsPanelItem, "adsr-curve"))
-        if (auto* curve = dynamic_cast<AdsrCurveComponent*>(item->getComponent().get()))
-            curve->setParameters(attack, decay, sustain, release);
-}
-
-void MainComponent::setControlsPlaybackSpeed(double speed) {
-    if (jiveControlsPanelItem == nullptr)
-        return;
-    if (auto* item = jive::findItemWithID(*jiveControlsPanelItem, "speed-knob"))
-        if (auto* slider = dynamic_cast<juce::Slider*>(item->getComponent().get()))
-            slider->setValue(juce::jlimit(0.5, 2.0, speed), juce::dontSendNotification);
-}
-
-void MainComponent::setControlsPresets(const juce::StringArray& presetIds, const juce::String& currentPresetId,
-                                       const juce::StringArray& presetDisplayNames) {
-    availablePresetIds = presetIds;
-    if (jiveControlsPanelItem == nullptr)
-        return;
-
-    auto* comboItem = jive::findItemWithID(*jiveControlsPanelItem, "preset-combo");
-    auto* combo = comboItem != nullptr ? dynamic_cast<juce::ComboBox*>(comboItem->getComponent().get()) : nullptr;
-
-    if (comboItem != nullptr)
-        comboItem->state.removeAllChildren(nullptr);
-    auto selectedIndex = 0;
-    for (int i = 0; i < presetIds.size(); ++i) {
-        const auto displayName = (i < presetDisplayNames.size()) ? presetDisplayNames[i] : presetIds[i];
-        if (comboItem != nullptr) {
-            auto option = juce::ValueTree("Option");
-            option.setProperty("text", displayName, nullptr);
-            comboItem->state.addChild(option, i, nullptr);
-        }
-        if (presetIds[i] == currentPresetId)
-            selectedIndex = i;
-    }
-    if (comboItem != nullptr)
-        comboItem->state.setProperty("selected", selectedIndex, nullptr);
-
-    updateControlsPresetActionButtons();
-}
-
-juce::String MainComponent::getSelectedPresetId() const {
-    if (jiveControlsPanelItem == nullptr)
-        return {};
-    if (auto* item = jive::findItemWithID(*jiveControlsPanelItem, "preset-combo"))
-        if (auto* combo = dynamic_cast<juce::ComboBox*>(item->getComponent().get())) {
-            const auto index = combo->getSelectedItemIndex();
-            if (juce::isPositiveAndBelow(index, availablePresetIds.size()))
-                return availablePresetIds[index];
-        }
-    return {};
-}
-
-void MainComponent::updateControlsPresetActionButtons() {
-    const auto isUserPreset = getSelectedPresetId().isNotEmpty();
-    if (jiveControlsPanelItem == nullptr)
-        return;
-    const auto setEnabled = [this](const char* id, bool enabled) {
-        if (auto* item = jive::findItemWithID(*jiveControlsPanelItem, id))
-            item->state.setProperty("enabled", enabled, nullptr);
-    };
-    setEnabled("rename-preset-btn", isUserPreset);
-    setEnabled("delete-preset-btn", isUserPreset);
-}
-
-void MainComponent::setRecordingControlsState(RecordingControlsState state) {
-    recordingControlsState = state;
-    if (jiveControlsPanelItem == nullptr)
-        return;
-
-    auto recordEnabled = true;
-    auto playEnabled = state.hasTake;
-    auto stopEnabled = false;
-    auto backToStartEnabled = state.hasTake;
-    auto exportMidiEnabled = state.hasTake && state.canExportMidiTake;
-    auto exportWavEnabled = state.hasTake && state.canExportWavTake;
-    auto importMidiEnabled = true;
-    auto saveEnabled = false;
-    auto openEnabled = false;
-
-    switch (state.state) {
-    case RecordingState::idle:
-        playEnabled = state.hasTake;
-        backToStartEnabled = state.hasTake;
-        exportMidiEnabled = state.hasTake && state.canExportMidiTake;
-        exportWavEnabled = state.hasTake && state.canExportWavTake;
-        importMidiEnabled = true;
-        saveEnabled = state.hasTake;
-        openEnabled = true;
-        break;
-    case RecordingState::recording:
-        recordEnabled = false;
-        playEnabled = false;
-        backToStartEnabled = false;
-        exportMidiEnabled = false;
-        exportWavEnabled = false;
-        importMidiEnabled = false;
-        stopEnabled = true;
-        saveEnabled = false;
-        openEnabled = false;
-        break;
-    case RecordingState::playing:
-        recordEnabled = false;
-        playEnabled = false;
-        backToStartEnabled = state.hasTake;
-        exportMidiEnabled = false;
-        exportWavEnabled = false;
-        importMidiEnabled = false;
-        stopEnabled = true;
-        saveEnabled = false;
-        openEnabled = false;
-        break;
-    }
-
-    const auto setEnabled = [this](const char* id, bool enabled) {
-        if (auto* item = jive::findItemWithID(*jiveControlsPanelItem, id))
-            item->state.setProperty("enabled", enabled, nullptr);
-    };
-    setEnabled("record-btn", recordEnabled);
-    setEnabled("play-btn", playEnabled);
-    setEnabled("stop-btn", stopEnabled);
-    setEnabled("back-btn", backToStartEnabled);
-    setEnabled("import-midi-btn", importMidiEnabled);
-    setEnabled("export-midi-btn", exportMidiEnabled);
-    setEnabled("export-wav-btn", exportWavEnabled);
-    setEnabled("save-perf-btn", saveEnabled);
-    setEnabled("open-perf-btn", openEnabled);
-}
-
-juce::Rectangle<int> MainComponent::getRecentFilesButtonScreenBounds() const {
-    if (jiveControlsPanelItem == nullptr)
-        return {};
-    if (auto* item = jive::findItemWithID(*jiveControlsPanelItem, "recent-btn"))
-        return item->getComponent()->getScreenBounds();
-    return {};
-}
-
-void MainComponent::refreshControlsTexts() {
-    if (jiveControlsPanelItem == nullptr)
-        return;
-
-    const auto setText = [this](const char* id, const juce::String& text) {
-        if (auto* item = jive::findItemWithID(*jiveControlsPanelItem, id))
-            item->state.setProperty("text", text, nullptr);
-    };
-    const auto setButtonText = [this](const char* id, const juce::String& text) {
-        if (auto* item = jive::findItemWithID(*jiveControlsPanelItem, id))
-            item->state.setProperty("text", text, nullptr);
-    };
-    const auto setTooltip = [this](const char* id, const juce::String& tooltip) {
-        if (auto* item = jive::findItemWithID(*jiveControlsPanelItem, id))
-            item->state.setProperty("tooltip", tooltip, nullptr);
-    };
-
-    setText("volume-label", TRANS("Volume"));
-    setText("attack-label", TRANS("Attack"));
-    setText("decay-label", TRANS("Decay"));
-    setText("sustain-label", TRANS("Sustain"));
-    setText("release-label", TRANS("Release"));
-    setText("preset-label", TRANS("Preset"));
-    setText("speed-label", TRANS("Speed"));
-    setButtonText("save-preset-btn", TRANS("Save As New"));
-    setButtonText("rename-preset-btn", TRANS("Rename"));
-    setButtonText("delete-preset-btn", TRANS("Delete"));
-    setTooltip("record-btn", TRANS("Record"));
-    setTooltip("play-btn", TRANS("Play"));
-    setTooltip("stop-btn", TRANS("Stop"));
-    setTooltip("back-btn", TRANS("Back to Start"));
-    setButtonText("export-midi-btn", TRANS("Export MIDI"));
-    setButtonText("export-wav-btn", TRANS("Export WAV"));
-    setButtonText("import-midi-btn", TRANS("Import MIDI"));
-    setButtonText("recent-btn", TRANS("Recent"));
-    setButtonText("save-perf-btn", TRANS("Save"));
-    setButtonText("open-perf-btn", TRANS("Open"));
-    setButtonText("song-info-btn", TRANS("Song Info"));
-
-    setRecordingControlsState(recordingControlsState);
-}
-
-// ── JIVE keyboard area accessors ───────────────────────────────────────────
-
-CustomKeyboard& MainComponent::getCustomKeyboard() {
-    jassert(customKeyboardRef != nullptr);
-    return *customKeyboardRef;
-}
-
-void MainComponent::setKeyboardLayout(const devpiano::core::KeyboardLayout& layout) {
-    if (jiveKeyboardAreaItem == nullptr)
-        return;
-    if (auto* item = jive::findItemWithID(*jiveKeyboardAreaItem, "custom-keyboard"))
-        if (auto* viewport = dynamic_cast<KeyboardViewport*>(item->getComponent().get()))
-            viewport->getCustomKeyboard().setKeyboardLayout(layout);
-}
-
-void MainComponent::setKeyboardViewPosition(int midiNote, int pixelOffset) {
-    if (jiveKeyboardAreaItem == nullptr)
-        return;
-    auto* item = jive::findItemWithID(*jiveKeyboardAreaItem, "custom-keyboard");
-    auto* viewport = item != nullptr ? dynamic_cast<KeyboardViewport*>(item->getComponent().get()) : nullptr;
-    if (viewport == nullptr)
-        return;
-
-    auto& keyboard = viewport->getCustomKeyboard();
-    if (pixelOffset >= 0) {
-        viewport->setViewPosition(pixelOffset, 0);
-    } else if (midiNote >= 0 && midiNote <= 127) {
-        int whiteCount = 0;
-        for (int n = 0; n < midiNote; ++n)
-            if (devpiano::ui::isWhiteKey(n))
-                ++whiteCount;
-        auto x = static_cast<int>(static_cast<float>(whiteCount) * keyboard.getKeyboardSettings().keyWidth);
-        viewport->setViewPosition(x, 0);
-    }
-}
-
-int MainComponent::getKeyboardViewPositionX() const noexcept {
-    if (jiveKeyboardAreaItem == nullptr)
-        return 0;
-    if (auto* item = jive::findItemWithID(*jiveKeyboardAreaItem, "custom-keyboard"))
-        if (auto* viewport = dynamic_cast<KeyboardViewport*>(item->getComponent().get()))
-            return viewport->getViewPositionX();
-    return 0;
-}
-
-void MainComponent::finishPluginUiAction(bool shouldSaveSettings) {
-    if (shouldSaveSettings)
-        saveSettingsSoon();
-
-    refreshReadOnlyUiStateFromCurrentSnapshot();
-    restoreKeyboardFocus();
-}
-
-void MainComponent::logCurrentAudioDeviceDiagnostics(const juce::String& context) const {
-    const auto diagnostics
-        = devpiano::audio::buildAudioDeviceDiagnostics(appSettings.audioDeviceState.get(), deviceManager);
-    DP_LOG_INFO("[AudioDevice] " + context + "\n" + diagnostics.detailedSummary);
-}
-
-devpiano::core::AppState MainComponent::buildCurrentAppStateSnapshot() const {
-    return devpiano::core::buildCurrentAppStateSnapshot(
-        appSettings, deviceManager, pluginHost,
-        pluginOperationController != nullptr && pluginOperationController->hasEditorWindowOpen(), keyboardMidiMapper);
-}
-
-void MainComponent::applyLanguage(const juce::String& code) {
-    devpiano::locale::activate(devpiano::locale::codeToLanguage(code));
-    refreshAllTexts();
-}
-
-void MainComponent::refreshAllTexts() {
-    refreshPluginPanelTexts();
-    refreshControlsTexts();
-}
-
-double MainComponent::getCurrentRuntimeSampleRate() const {
-    if (auto* device = deviceManager.getCurrentAudioDevice()) {
-        const auto rate = device->getCurrentSampleRate();
-        if (rate > 0.0)
-            return rate;
-    }
-
-    return appSettings.getAudioSettingsView().sampleRate;
-}
-
-int MainComponent::getCurrentRuntimeBlockSize() const {
-    if (auto* device = deviceManager.getCurrentAudioDevice()) {
-        const auto size = device->getCurrentBufferSizeSamples();
-        if (size > 0)
-            return size;
-    }
-
-    return appSettings.getAudioSettingsView().bufferSize;
-}
-
-MainComponent::RuntimeAudioConfig MainComponent::getCurrentRuntimeAudioConfig() const {
-    return { .sampleRate = getCurrentRuntimeSampleRate(), .blockSize = getCurrentRuntimeBlockSize() };
-}
-
-void MainComponent::runPluginActionWithAudioDeviceRebuild(
-    const std::function<void(const RuntimeAudioConfig&)>& action) {
-    struct AudioDeviceRebuildGuard final {
-        explicit AudioDeviceRebuildGuard(MainComponent& ownerIn)
-            : owner(ownerIn) {
-        }
-        ~AudioDeviceRebuildGuard() {
-            owner.finishAudioDeviceRebuild();
-        }
-
-        MainComponent& owner;
-    };
-
-    const auto runtimeAudioConfig = getCurrentRuntimeAudioConfig();
-
-    prepareForAudioDeviceRebuild();
-    const AudioDeviceRebuildGuard rebuildGuard(*this);
-    action(runtimeAudioConfig);
-}
-
-void MainComponent::runPluginActionWithAudioDeviceRebuild(const std::function<void()>& action) {
-    runPluginActionWithAudioDeviceRebuild([&action](const RuntimeAudioConfig&) { action(); });
-}
-
-void MainComponent::saveRecentFiles() {
-    appSettings.recentFilesSerialized = recentFiles.toString();
-    saveSettingsSoon();
-}
-
-void MainComponent::showRecentFilesMenu() {
-    juce::PopupMenu menu;
-    recentFiles.removeNonExistentFiles();
-
-    const auto numFiles = recentFiles.getNumFiles();
-    int itemId = 1;
-
-    if (numFiles == 0) {
-        menu.addItem(0, TRANS("(no recent files)"), false, false);
-    } else {
-        for (int i = 0; i < numFiles; ++i) {
-            auto file = recentFiles.getFile(i);
-            auto name = file.getFileName();
-            auto ext = file.getFileExtension().toLowerCase();
-            juce::String prefix;
-            if (ext == ".devpiano")
-                prefix = juce::String::fromUTF8("\xe2\x99\xaa "); // ♪
-            else if (ext == ".mid" || ext == ".midi")
-                prefix = juce::String::fromUTF8("\xe2\x99\xab "); // ♫
-            else
-                prefix = "? ";
-
-            menu.addItem(itemId, prefix + name);
-            ++itemId;
-        }
-    }
-
-    int clearId = itemId;
-    if (numFiles > 0) {
-        menu.addSeparator();
-        clearId = itemId;
-        menu.addItem(clearId, TRANS("Clear Recent Files"));
-    }
-
-    menu.showMenuAsync(juce::PopupMenu::Options().withTargetScreenArea(getRecentFilesButtonScreenBounds()),
-                       [safe = juce::Component::SafePointer<MainComponent>(this), numFiles, clearId](int result) {
-                           if (safe == nullptr)
-                               return;
-
-                           if (result == 0)
-                               return;
-
-                           if (result == clearId) {
-                               safe->recentFiles.clear();
-                               safe->saveRecentFiles();
-                               return;
-                           }
-
-                           const auto index = result - 1;
-                           if (!juce::isPositiveAndBelow(index, numFiles))
-                               return;
-
-                           auto file = safe->recentFiles.getFile(index);
-                           if (!file.exists())
-                               return;
-
-                           auto ext = file.getFileExtension().toLowerCase();
-                           if (ext == ".devpiano")
-                               safe->recordingSessionController->handleOpenPerformanceFile(file);
-                           else if (ext == ".mid" || ext == ".midi")
-                               safe->recordingSessionController->handleImportMidiFile(file);
-                       });
-}
+// JIVE component accessors (see MainComponentJiveAccessors.cpp)
+#include "MainComponentJiveAccessors.cpp"
